backend/django_backend/requirements.txt
Django
djangorestframework
psycopg2
newspaper3k
pandas
sklearn